Delete file
Realised online Jupyter Notebook Viewer does not take the latest version (latest commit) of a file from a github repo, only the original version of the file that was uploaded. The original version, in my case, has the wrong path (local instead of github url), and so I decided to just reupload a new file.